Added orders, exceptions, some user functionality too... #28

+++ src/main/java/com/sharashkina_kontora/travel_agency/service/FlightServiceImpl.java
@@ -1,6 +1,7 @@
 package com.sharashkina_kontora.travel_agency.service;
 
 import com.sharashkina_kontora.travel_agency.domain.Flight;
+import com.sharashkina_kontora.travel_agency.exceptions.FlightInUseException;
 import com.sharashkina_kontora.travel_agency.repository.FlightRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -59,7 +60,8 @@ public class FlightServiceImpl implements FlightService {
      */
     @Override
     @Transactional
-    public void delete(Flight flight) {
+    public void delete(Flight flight) throws Exception {
+        if (!flight.getTours().isEmpty()) throw new FlightInUseException("Flight in use");
         flightRepository.delete(flight);
     }
